+++ .gitignore
@@ -8,9 +8,8 @@ __pycache__/
 *.db
 *.whl
 pc_transformer_venv/
-data_preprocessing/
 
 venv/
 *.egg-info/
 checkpoints/
-Data_preprocessing/Data/opwb/
+data_preparation/data/opwb/
